module: Reconfiguration of modules (single module -> multi module)
