+++ fluxgate-core/src/main/java/org/fluxgate/core/key/KeyResolver.java
@@ -1,0 +1,7 @@
+package org.fluxgate.core.key;
+
+import org.fluxgate.core.context.RequestContext;
+
+public interface KeyResolver {
+    RateLimitKey resolve(RequestContext context);
+}
