--- src/main/java/org/openfluxgate/core/key/KeyResolver.java
@@ -1,7 +1,0 @@
-package org.openfluxgate.core.key;
-
-import org.openfluxgate.core.context.RequestContext;
-
-public interface KeyResolver {
-    RateLimitKey resolve(RequestContext context);
-}
